removed type checking for advanveTurn
