+++ source/battleActions/BattlePhase.h
@@ -1,0 +1,12 @@
+#pragma once
+
+enum class BattlePhase
+{
+    ChooseAction,
+    ChooseSkill,
+    ChoosePersona,
+    ChooseTarget,
+    ShowAlert,
+    EnemyTurn,
+    Done
+};
