+++ source/battleActions/party/PartyMember.h
@@ -31,6 +31,7 @@ struct PartyMember : BattleParticipant
     float calculateBaseDamage(BattleParticipant& defender, Skill& skill) override;
     float getTeamMultiplier() override;
     void setCurrentTurnOrderAgility(float boost) override;
+    BattlePhase getInitalTurnPhase() override;
     virtual bool actorCanUse(ActionBase* action);
 
     ~PartyMember()
